Create basic "command framework".
Each command is represented by a class inheriting BotCommand. Each command must overwrite
BotCommand.parseMessage() method which returns a reply (or None when there is no reply). Reply is
simply a string where each line (ended by \n) represents a single PRIVMSG. Additionaly commands
provide a little help in form of docstring which is automatically parsed and returned by a
HelpCommand. Commands must be registered in a teamercommands.commands dictionary.

One of the simplest commands is 'ShufleCommand', so take a look at it if you want an easy
introduction.

Additionaly SET and WHOIS command werre removed because they were useless... SHUFFLE and LIST are
added instead.